Move goal evaluation functions
